\name{CRD1}
\alias{CRD1}

\title{
  Completely Randomized Design (CRD)
}

\description{
  A \code{list} illustrating the resources of \pkg{ScottKnott} package
  related to Completely Randomized Design (\samp{CRD}).
}

\usage{
  data(CRD1)
  CRD1
}

\details{
  Simulated data for a Completely Randomized Design (\samp{CRD}) with 4
  treatment levels and 6 replicates per treatment.
}

\keyword{datasets}
