\name{CRD2}
\alias{CRD2}

\title{
  Completely Randomized Design (CRD)
}

\description{
  A \code{list} illustrating the resources of \pkg{ScottKnott} package
  related to Completely Randomized Design (\samp{CRD}).
}

\usage{
  data(CRD2)
  CRD2
}

\details{
  Simulated data for a Completely Randomized Design (\samp{CRD}) with 45
  treatment levels and 4 replicates per treatment.
}

\keyword{datasets}
