\name{FE}
\alias{FE}

\title{Factorial Experiment (FE)}

\description{
  A \code{list} illustrating  the resources of \code{ScottKnott} package
  related to Factorial Experiment (FE).
}

\usage{
  data(FE)
  FE
}

\details{
  A simulated data to model a Factorial Experiment (FE) with 3 factors, 
  2 levels per factor and 4 blocks.
}

\keyword{datasets}
