\name{LSD}
\alias{LSD}

\title{
  Latin Squares Design (LSD)
}

\description{
  A \code{list} illustrating the resources of \pkg{ScottKnott} package
  related to Latin Squares Design (\samp{LSD}).
}

\usage{
  data(LSD)
  LSD
}

\details{
  Simulated data for a Latin Squares Design (\samp{LSD}) with 5 treatment
  levels, 5 rows, and 5 columns.
}

\keyword{datasets}
