\name{RCBD}
\alias{RCBD}

\title{
  Randomized Complete Block Design (RCBD)
}

\description{
  A \code{list} illustrating the resources of \pkg{ScottKnott} package
  related to Randomized Complete Block Design (\samp{RCBD}).
}

\usage{
  data(RCBD)
  RCBD
}

\details{
  Simulated data for a Randomized Complete Block Design (\samp{RCBD}) with
  5 treatment levels and 4 blocks (one replicate per block).
}

\keyword{datasets}
