\name{SK-internal}
\title{Internal SK functions}
\alias{m.infos.aovlist}
\alias{m.infos.lm}
\alias{m.infos.lmerMod}
\alias{m.infos.nest.aovlist}
\alias{m.infos.nest.lm}
\alias{m.infos.nest.lmerMod}
\alias{MaxValue}
\description{Internal SK functions.}
\details{These are not to be called by the user and are undocumented.}
\keyword{internal}
