\name{SPE}
\alias{SPE}

\title{
  Split-Plot Experiment (SPE)
}

\description{
  A \code{list} illustrating the resources of \pkg{ScottKnott} package
  related to Split-Plot Experiment (\samp{SPE}).
}

\usage{
  data(SPE)
  SPE
}

\details{
  Simulated data for a Split-Plot Experiment (\samp{SPE}) with 3 whole plots,
  4 sub-plot treatments, and 6 replicates per sub-plot.
}

\keyword{datasets}
