\name{SSPE}
\alias{SSPE}

\title{
  Split-Split-Plot Experiment (SSPE)
}

\description{
  A \code{list} illustrating the resources of \pkg{ScottKnott} package
  related to Split-Split-Plot Experiment (\samp{SSPE}).
}

\usage{
  data(SSPE)
  SSPE
}

\details{
  Simulated data for a Split-Split-Plot Experiment (\samp{SSPE}) with 3 whole
  plots, 3 sub-plot treatments, 5 sub-sub-plot treatments, and 4 replicates
  per sub-sub-plot.
}

\keyword{datasets}
