\name{xtable}
\alias{xtable}

\title{Create a Table for Export}

\description{
  This function is re-exported from the \pkg{xtable} package so that
  \code{xtable()} is available after \code{library(ScottKnott)} without
  requiring a separate \code{library(xtable)} call.

  For \code{SK} objects the S3 method \code{\link{xtable.SK}} is dispatched
  automatically.  For full documentation of the generic see
  \code{help("xtable", package = "xtable")}.
}

\seealso{
  \code{\link{xtable.SK}}, \code{\link[xtable]{xtable}}
}

\keyword{utilities}
