\name{CRD2}
\alias{CRD2}

\title{
  Completely Randomized Design (CRD)
}

\description{
  A \code{list} illustrating the resources of \pkg{ScottKnott} package
  related to Completely Randomized Design (\samp{CRD}).
}

\usage{
  data(CRD2)
  CRD2
}

\details{
  A simulated data to model a Completely Randomized Design (\samp{CRD}) of 45 
  factor levels and 4 repetitions.
}

\keyword{datasets}
