\name{LSD}
\alias{LSD}

\title{Latin Squares Design (LSD)}

\description{
  A \code{list} illustrating  the resources of \code{ScottKnott} package
  related to Latin Squares Design (LSD).
}

\usage{
  data(LSD)
  LSD
}

\details{
  A simulated data to model a Latin Squares Design (LSD) with  5
  factor levels 5 rows and 5 columns.
}

\keyword{datasets}
